added test setup

+++ src/main/resources/db/migration/V2__employee.sql
@@ -5,7 +5,7 @@ CREATE SEQUENCE employee_id_seq START 1000;
 CREATE TABLE IF NOT EXISTS employee
 (
 
-    id           INTEGER DEFAULT nextval('employee_id_seq') PRIMARY KEY,
+    id           BIGINT DEFAULT nextval('employee_id_seq') PRIMARY KEY,
     first_name   VARCHAR(255) NOT NULL,
     last_name    VARCHAR(255) NOT NULL,
     abbreviation VARCHAR(255) NOT NULL UNIQUE
